backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "astroid==3.3.10",
    "autopep8==2.0.4",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.2.1",
    "cryptography==45.0.3",
    "dill==0.4.0",
    "dnspython==2.7.0",
    "docstring-to-markdown==0.17",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "flake8==7.1.2",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "isort==6.0.1",
    "jedi==0.19.2",
    "mccabe==0.7.0",
    "minio>=7.2.15",
    "mypy==1.16.1",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "parso==0.8.4",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "psycopg2==2.9.10",
    "pyasn1==0.6.1",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydocstyle==6.3.0",
    "pyflakes==3.2.0",
    "pylint==3.3.7",
    "pylsp-mypy==0.7.0",
    "pylsp-rope==0.1.17",
    "pyotp==2.9.0",
    "python-dotenv==1.1.0",
    "python-jose==3.5.0",
    "python-lsp-jsonrpc==1.1.2",
    "python-lsp-server==1.12.2",
    "python-multipart==0.0.20",
    "pytoolconfig==1.3.1",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "rope==1.13.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowballstemmer==3.0.1",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "tomlkit==0.13.3",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "ujson==5.10.0",
    "uvicorn==0.34.3",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "websockets==15.0.1",
    "whatthepatch==1.0.7",
    "yapf==0.43.0",
    "zipp==3.23.0",
]

[dependency-groups]
dev = [
    "nox>=2025.5.1",
    "pyrefly>=0.21.0",
    "ruff>=0.12.0",
]
